+++ example/windows/example/pch.h
@@ -6,16 +6,12 @@
 #include <restrictederrorinfo.h>
 #include <unknwn.h>
 #include <windows.h>
+#include <CppWinRTIncludes.h>
 #include <winrt/Windows.ApplicationModel.Activation.h>
-#include <winrt/Windows.Foundation.Collections.h>
-#include <winrt/Windows.Foundation.h>
-#include <winrt/Windows.UI.Xaml.Controls.Primitives.h>
-#include <winrt/Windows.UI.Xaml.Controls.h>
-#include <winrt/Windows.UI.Xaml.Data.h>
-#include <winrt/Windows.UI.Xaml.Interop.h>
-#include <winrt/Windows.UI.Xaml.Markup.h>
-#include <winrt/Windows.UI.Xaml.Navigation.h>
-#include <winrt/Windows.UI.Xaml.h>
+#include <UI.Xaml.Controls.Primitives.h>
+#include <UI.Xaml.Controls.h>
+#include <UI.Xaml.Markup.h>
+#include <UI.Xaml.Navigation.h>
 
 #include <winrt/Microsoft.ReactNative.h>